+++ framework-system/framework-system-persistence/src/main/java/uk/gov/justice/services/system/persistence/SystemPersistenceException.java
@@ -1,0 +1,12 @@
+package uk.gov.justice.services.system.persistence;
+
+public class SystemPersistenceException extends RuntimeException {
+
+    public SystemPersistenceException(final String message) {
+        super(message);
+    }
+
+    public SystemPersistenceException(final String message, final Throwable cause) {
+        super(message, cause);
+    }
+}